+++ src/components/Menu.module.css
@@ -6,7 +6,7 @@
   z-index: 1000;
   width: 100%;
   background: #fff;
-  padding-top: 100px;
+  padding: 100px 1rem;
   transform: translateX(-100%);
   will-change: transform;
 }
@@ -19,7 +19,7 @@ ul {
   justify-content: space-between;
 }
 
-li {
+.menu-item {
   color: #000;
   display: block;
   transform-origin: -20px 50%;
@@ -28,7 +28,6 @@ li {
   padding: 10px;
   will-change: transform, opacity, filter;
   flex-basis: 30%;
-  /* background-color: red; */
   text-align: center;
   text-transform: uppercase;
   position: relative;
@@ -61,4 +60,8 @@ img {
   width: 100%;
   object-fit: cover;
   filter: grayscale(100%);
+}
+
+.menu-item:hover {
+  box-shadow: 0 2px rgba(0, 0, 0, 0.2);
 }
