+++ src/index.css
@@ -15,72 +15,3 @@
   padding: 0;
   box-sizing: border-box;
 }
-
-.container {
-  padding: 2rem 2rem 0;
-  overflow-y: hidden;
-}
-
-.top-layer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  position: relative;
-
-  img {
-    width: 65%;
-    filter: grayscale();
-  }
-  .box {
-    width: 315px;
-    height: 100px;
-    background-color: #fff;
-    position: absolute;
-    bottom: 20px;
-    left: 33%;
-  }
-}
-
-.title {
-  font-size: 446px;
-  font-weight: bold;
-  margin: 0;
-  line-height: 1;
-}
-
-.bottom-layer {
-  .content {
-    display: flex;
-    position: relative;
-    justify-content: space-between;
-    align-items: center;
-  }
-  .title {
-    margin-right: 140px;
-  }
-  .vertical {
-    position: absolute;
-    transform: rotate(90deg);
-    font-size: 74px;
-    bottom: 170px;
-    left: 780px;
-  }
-  .personal-info {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-end;
-    height: 380px;
-    font-size: 32px;
-  }
-  .line {
-    display: flex;
-    justify-content: space-between;
-    border-bottom: 1px solid #000;
-    padding: 1rem 0;
-  }
-  .link {
-    color: inherit;
-    text-decoration: none;
-  }
-}
