+++ src/pages/RootLayout/RootLayout.module.css
@@ -1,0 +1,4 @@
+.container {
+  padding: 0 1rem;
+  overflow-y: hidden;
+}
