+++ src/pages/WelcomePage/WelcomePage.module.css
@@ -1,0 +1,61 @@
+.top-layer {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  position: relative;
+
+  img {
+    width: 65%;
+    filter: grayscale();
+  }
+  .box {
+    width: 315px;
+    height: 160px;
+    background-color: #fff;
+    position: absolute;
+    bottom: 10px;
+    left: 33%;
+  }
+}
+
+.title {
+  font-size: 446px;
+  font-weight: bold;
+  margin: 0;
+  line-height: 1;
+}
+
+.bottom-layer {
+  .content {
+    display: flex;
+    position: relative;
+    justify-content: space-between;
+    align-items: center;
+  }
+  .title {
+    margin-right: 140px;
+  }
+
+  .personal-info {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-end;
+    height: 380px;
+    font-size: 32px;
+  }
+  .line {
+    display: flex;
+    justify-content: space-between;
+    border-bottom: 1px solid #000;
+    padding: 1rem 0;
+  }
+}
+
+.vertical {
+  position: absolute;
+  transform: rotate(90deg);
+  font-size: 74px;
+  bottom: 170px;
+  left: 780px;
+}
